feat: v0.4: Reduced Webview files for bundle size + Delete on Ingest Functionailty
