+++ src/webview/index.ts
@@ -1,0 +1,96 @@
+import { THEME } from '../config';
+import { ButtonProps, SectionProps, StatusMessage, ThemeColors } from '../types';
+
+// Helpers
+
+function escapeHtml(text: string): string {
+    return text
+        .replace(/&/g, '&amp;')
+        .replace(/</g, '&lt;')
+        .replace(/>/g, '&gt;')
+        .replace(/"/g, '&quot;')
+        .replace(/'/g, '&#039;');
+}
+
+const icons = {
+    copy: '<path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M8 5H6a2 2 0 00-2 2v12a2 2 0 002 2h10a2 2 0 002-2v-1M8 5a2 2 0 002 2h2a2 2 0 002-2M8 5a2 2 0 012-2h2a2 2 0 012 2m0 0h2a2 2 0 012 2v3m2 4H10m0 0l3-3m-3 3l3 3" />',
+    save: '<path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M8 7H5a2 2 0 00-2 2v9a2 2 0 002 2h14a2 2 0 002-2V9a2 2 0 00-2-2h-3m-1 4l-3 3m0 0l-3-3m3 3V4" />',
+    setup: '<path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M10.325 4.317c.426-1.756 2.924-1.756 3.35 0a1.724 1.724 0 002.573 1.066c1.543-.94 3.31.826 2.37 2.37a1.724 1.724 0 001.065 2.572c1.756.426 1.756 2.924 0 3.35a1.724 1.724 0 00-1.066 2.573c.94 1.543-.826 3.31-2.37 2.37a1.724 1.724 0 00-2.572 1.065c-.426 1.756-2.924 1.756-3.35 0a1.724 1.724 0 00-2.573-1.066c-1.543.94-3.31-.826-2.37-2.37a1.724 1.724 0 00-1.065-2.572c-1.756-.426-1.756-2.924 0-3.35a1.724 1.724 0 001.066-2.573c-.94-1.543.826-3.31 2.37-2.37.996.608 2.296.07 2.572-1.065z" /><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M15 12a3 3 0 11-6 0 3 3 0 016 0z" />',
+    retry: '<path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M4 4v5h.582m15.356 2A8.001 8.001 0 004.582 9m0 0H9m11 11v-5h-.581m0 0a8.003 8.003 0 01-15.357-2m15.357 2H15" />',
+    play: '<path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M5 3l14 9-14 9V3z" />',
+    success:
+        '<path d="M22 11.08V12a10 10 0 1 1-5.93-9.14"></path><polyline points="22 4 12 14.01 9 11.01"></polyline>',
+    error: '<circle cx="12" cy="12" r="10"></circle><line x1="15" y1="9" x2="9" y2="15"></line><line x1="9" y1="9" x2="15" y2="15"></line>',
+    warning:
+        '<path d="M10.29 3.86L1.82 18a2 2 0 0 0 1.71 3h16.94a2 2 0 0 0 1.71-3L13.71 3.86a2 2 0 0 0-3.42 0z"></path><line x1="12" y1="9" x2="12" y2="13"></line><line x1="12" y1="17" x2="12.01" y2="17"></line>',
+    info: '<circle cx="12" cy="12" r="10"></circle><line x1="12" y1="16" x2="12" y2="12"></line><line x1="12" y1="8" x2="12.01" y2="8"></line>',
+};
+
+function getStatusIcon(type: string): string {
+    switch (type) {
+        case 'success':
+            return `<svg width="20" height="20" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round">${icons.success}</svg>`;
+        case 'error':
+            return `<svg width="20" height="20" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round">${icons.error}</svg>`;
+        case 'warning':
+            return `<svg width="20" height="20" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round">${icons.warning}</svg>`;
+        default:
+            return `<svg width="20" height="20" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round">${icons.info}</svg>`;
+    }
+}
+
+// Components
+
+function Button({ onClick, variant = 'primary', icon, children }: ButtonProps): string {
+    const variantClasses = { primary: 'primary-button', danger: 'danger-button' };
+    return `<button class="button ${variantClasses[variant]}" onclick="${onClick}">${icon ? `<svg width="16" height="16" fill="none" stroke="currentColor" viewBox="0 0 24 24">${icon}</svg>` : ''}${children}</button>`;
+}
+
+function Section({ title, content, copyButton = true, copyFunction }: SectionProps): string {
+    return `<div><div class="section-header"><h3 class="section-title">${title}</h3>${copyButton ? `<button class="primary-button" onclick="${copyFunction || ''}"><svg width="16" height="16" fill="none" stroke="currentColor" viewBox="0 0 24 24"><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M8 5H6a2 2 0 00-2 2v12a2 2 0 002 2h10a2 2 0 002-2v-1M8 5a2 2 0 002 2h2a2 2 0 002-2M8 5a2 2 0 012-2h2a2 2 0 012 2m0 0h2a2 2 0 012 2v3m2 4H10m0 0l3-3m-3 3l3 3"/></svg>Copy</button>` : ''}</div><div class="section-shadow-wrapper"><div class="content-box"><div class="scrollable-content"><pre>${content}</pre></div></div></div></div>`;
+}
+
+// Styles
+
+const getBaseStyles = (theme: ThemeColors = THEME) => `
+body {font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, sans-serif;padding: 20px;color: #1a1a1a;line-height: 1.6;background-color: #FFFDF8;}.shadow-wrapper {position: relative;}.shadow-wrapper::before {content: '';position: absolute;inset: 0;background: #1a1a1a;border-radius: 12px;transform: translate(2px, 2px);z-index: 10;}.content-wrapper {background: #fafafa;border: 3px solid #1a1a1a;border-radius: 12px;padding: 24px;position: relative;z-index: 20;}.loading-container {display: flex;justify-content: center;align-items: center;min-height: 400px;}.loading-content {width: 100%;max-width: 500px;text-align: center;}.loader-wrapper {margin-bottom: 2rem;}.loading-title {font-size: 1.5rem;font-weight: 600;margin: 1rem 0;color: #1a1a1a;}.loader {border: 4px solid #fff4da;border-top: 4px solid #ffc480;border-radius: 50%;width: 60px;height: 60px;animation: spin 1s linear infinite;margin: 0 auto;}.status-container {margin: 2rem 0;}.status-item {display: flex;align-items: center;justify-content: center;gap: 0.75rem;padding: 0.75rem;margin-bottom: 0.75rem;border-radius: 8px;background: #fff4da;border: 2px solid #1a1a1a;font-size: 1rem;transition: transform 0.2s ease;}.status-item:hover {transform: translateY(-1px);}.status-item.success {color: #27ae60;}.status-item.error {color: ${theme.danger};}.status-item.warning {color: #f39c12;}.status-item.info {color: #3498db;}.status-icon {display: flex;align-items: center;}.status-text {font-weight: 500;}@keyframes spin {0% {transform: rotate(0deg);}100% {transform: rotate(360deg);}}.inner-content {background: #fff4da;border: 3px solid #1a1a1a;border-radius: 12px;padding: 24px;position: relative;}.grid {display: grid;grid-template-columns: 1fr 1fr;gap: 24px;margin-bottom: 24px;}@media (max-width: 768px) {.grid {grid-template-columns: 1fr;}}.section-title {font-size: 1.25rem;font-weight: bold;color: #1a1a1a;margin-bottom: 16px;}.section-header {display: flex;justify-content: space-between;align-items: center;margin-bottom: 16px;}button {padding: 12px 24px;cursor: pointer;border-radius: 8px;font-weight: 600;font-size: 1rem;transition: all 0.2s;position: relative;z-index: 20;display: inline-flex;align-items: center;gap: 0.5rem;}.primary-button {background-color: #ffc480;color: #1a1a1a;border: 3px solid #1a1a1a;}.primary-button:hover {transform: translate(-1px, -1px);}.danger-button {background-color: ${theme.danger};color: white;border: 3px solid #1a1a1a;}.danger-button:hover {transform: translate(-1px, -1px);}.button-group {display: flex;gap: 12px;margin-top: 16px;}.section-shadow-wrapper {position: relative;margin-bottom: 16px;}.section-shadow-wrapper::before {content: '';position: absolute;inset: 0;background: #1a1a1a;border-radius: 8px;transform: translate(2px, 2px);z-index: 10;}.content-box {background: #fff4da;border: 3px solid #1a1a1a;border-radius: 8px;padding: 16px;position: relative;z-index: 20;display: flex;flex-direction: column;height: 100%;}textarea, pre {font-family: ui-monospace, SFMono-Regular, Menlo, Monaco, Consolas, monospace;font-size: 0.875rem;line-height: 1.5;padding: 12px;background: #fff4da;border: 3px solid #1a1a1a;border-radius: 4px;width: 100%;min-height: 150px;resize: vertical;white-space: pre-wrap;word-wrap: break-word;margin: 0;}.scrollable-content {position: relative;max-height: 300px;overflow-y: auto;overflow-x: hidden;border: 3px solid #1a1a1a;border-radius: 4px;background: #fff4da;}.scrollable-content pre {border: none;margin: 0;height: 100%;}.error-container {display: flex;justify-content: center;align-items: center;min-height: 400px;}.error-content {width: 100%;max-width: 500px;text-align: center;}.error-icon {color: ${theme.danger};margin-bottom: 1.5rem;}.error-title {font-size: 1.75rem;font-weight: 600;color: ${theme.danger};margin-bottom: 1.5rem;}.error-messages {margin: 2rem 0;}.error-message {display: flex;align-items: center;justify-content: center;gap: 0.75rem;padding: 0.75rem;margin-bottom: 0.75rem;border-radius: 8px;background: #fff4da;border: 2px solid #1a1a1a;font-size: 1rem;transition: transform 0.2s ease;}.error-message:hover {transform: translateY(-1px);}.error-message-icon {display: flex;align-items: center;color: ${theme.danger};}.error-message-text {font-weight: 500;color: #1a1a1a;}.error-actions {display: flex;gap: 1rem;justify-content: center;margin-top: 2rem;}
+`;
+
+// Templates
+
+export function getErrorContent(
+    title: string,
+    messages: string[],
+    showSetupButton: boolean = true,
+): string {
+    const content = `<div class="shadow-wrapper"><div class="content-wrapper"><div class="error-container"><div class="error-content"><div class="error-icon"><svg width="64" height="64" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><circle cx="12" cy="12" r="10"></circle><line x1="15" y1="9" x2="9" y2="15"></line><line x1="9" y1="9" x2="15" y2="15"></line></svg></div><h2 class="error-title">${title}</h2><div class="error-messages">${messages.map((msg) => `<div class="error-message"><span class="error-message-icon"><svg width="20" height="20" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><circle cx="12" cy="12" r="10"></circle><line x1="12" y1="8" x2="12" y2="12"></line><line x1="12" y1="16" x2="12.01" y2="16"></line></svg></span><span class="error-message-text">${msg}</span></div>`).join('')}</div>${showSetupButton ? `<div class="error-actions">${Button({ onClick: 'openSetupGuide()', children: 'Show Setup Guide', icon: icons.setup })}${Button({ onClick: 'retryAnalysis()', variant: 'primary', children: 'Retry Analysis', icon: icons.retry })}</div>` : ''}</div></div></div></div><script>function openSetupGuide(){vscode.postMessage({command:'showSetup'})}function retryAnalysis(){vscode.postMessage({command:'retry'})}</script>`;
+    return createHtmlDocument(content, getBaseStyles(THEME));
+}
+
+export function getLoadingContent(statusMessages: StatusMessage[]): string {
+    const content = `<div class="shadow-wrapper"><div class="content-wrapper"><div class="loading-container"><div class="loading-content"><div class="loader-wrapper"><div class="loader"></div><h2 class="loading-title">Analyzing Repository</h2></div><div class="status-container">${statusMessages.map((msg) => `<div class="status-item ${msg.type}"><span class="status-icon">${getStatusIcon(msg.type)}</span><span class="status-text">${msg.text}</span></div>`).join('')}</div>${Button({ onClick: 'cancelAnalysis()', variant: 'danger', children: 'Cancel Analysis' })}</div></div></div></div><script>function cancelAnalysis(){vscode.postMessage({command:'cancel'})}</script>`;
+    return createHtmlDocument(content, getBaseStyles(THEME));
+}
+
+export function getResultsContent(data: {
+    summary: string;
+    tree: string;
+    content: string;
+}): string {
+    const content = `<div class="shadow-wrapper"><div class="content-wrapper"><div class="grid"><div>${Section({ title: 'Summary', content: escapeHtml(data.summary), copyFunction: 'copySummary()' })}<div class="button-group">${Button({ onClick: 'copyAll()', icon: icons.copy, children: 'Copy All' })}${Button({ onClick: 'saveToFile()', icon: icons.save, children: 'Save to File' })}</div></div>${Section({ title: 'Directory Structure', content: escapeHtml(data.tree), copyFunction: 'copyTree()' })}</div>${Section({ title: 'Files Content', content: escapeHtml(data.content), copyFunction: 'copyContent()' })}</div></div><script>function copySummary(){const summary=document.querySelectorAll('.scrollable-content pre')[0].innerText;vscode.postMessage({command:'copy',text:summary})}function copyTree(){const tree=document.querySelectorAll('.scrollable-content pre')[1].innerText;vscode.postMessage({command:'copy',text:tree})}function copyContent(){const content=document.querySelectorAll('.scrollable-content pre')[2].innerText;vscode.postMessage({command:'copy',text:content})}function copyAll(){const allContent=[document.querySelectorAll('.scrollable-content pre')[0].innerText,document.querySelectorAll('.scrollable-content pre')[1].innerText,document.querySelectorAll('.scrollable-content pre')[2].innerText].join('\\n\\n');vscode.postMessage({command:'copy',text:allContent})}function saveToFile(){const summary=document.querySelectorAll('.scrollable-content pre')[0].innerText;const tree=document.querySelectorAll('.scrollable-content pre')[1].innerText;const content=document.querySelectorAll('.scrollable-content pre')[2].innerText;vscode.postMessage({command:'saveToFile',data:{summary,tree,content}})}</script>`;
+    return createHtmlDocument(content, getBaseStyles(THEME));
+}
+
+export function getSetupGuideContent(): string {
+    const content = `<div class="shadow-wrapper"><div class="content-wrapper"><div class="setup-guide"><div class="setup-header"><div class="setup-icon"><svg width="48" height="48" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><path d="M10.325 4.317c.426-1.756 2.924-1.756 3.35 0a1.724 1.724 0 002.573 1.066c1.543-.94 3.31.826 2.37 2.37a1.724 1.724 0 001.065 2.572c1.756.426 1.756 2.924 0 3.35a1.724 1.724 0 00-1.066 2.573c.94 1.543-.826 3.31-2.37 2.37a1.724 1.724 0 00-2.572 1.065c-.426 1.756-2.924 1.756-3.35 0a1.724 1.724 0 00-2.573-1.066c-1.543.94-3.31-.826-2.37-2.37a1.724 1.724 0 00-1.065-2.572c-1.756-.426-1.756-2.924 0-3.35a1.724 1.724 0 001.066-2.573c-.94-1.543.826-3.31 2.37-2.37.996.608 2.296.07 2.572-1.065z"/><path d="M15 12a3 3 0 11-6 0 3 3 0 016 0z"/></svg></div><h1 class="setup-title">GitIngest Setup Guide</h1><p class="setup-subtitle">Follow these steps to get started with GitIngest</p></div><div class="setup-steps"><div class="setup-step"><div class="step-number">1</div><div class="step-content"><h2 class="step-title">Install Python</h2><div class="step-description"><p>First, make sure you have Python installed on your system:</p><div class="step-card"><div class="card-item"><div class="card-icon"><svg width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><path d="M21 12V7a2 2 0 0 0-2-2H5a2 2 0 0 0-2 2v5"/><path d="M3 12v5a2 2 0 0 0 2 2h14a2 2 0 0 0 2-2v-5"/><path d="M12 12v8"/><path d="M12 12 8 8"/><path d="m12 12 4-4"/></svg></div><div class="card-text"><a href="https://python.org" class="card-link">Download Python from python.org</a></div></div><div class="card-item"><div class="card-icon"><svg width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><polyline points="20 6 9 17 4 12"/></svg></div><div class="card-text">Check "Add Python to PATH" during installation</div></div><div class="card-item"><div class="card-icon"><svg width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><polyline points="22 7 13.5 15.5 8.5 10.5 2 17"/><polyline points="16 7 22 7 22 13"/></svg></div><div class="card-text">Verify by running: <code class="inline-code">python --version</code> or <code class="inline-code">python3 --version</code></div></div></div></div></div></div><div class="setup-step"><div class="step-number">2</div><div class="step-content"><h2 class="step-title">Usage</h2><div class="step-description"><p>After installation, you're ready to analyze repositories:</p><div class="usage-steps"><div class="usage-step"><div class="usage-icon"><svg width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><rect x="3" y="3" width="18" height="18" rx="2" ry="2"/><line x1="3" y1="9" x2="21" y2="9"/><line x1="9" y1="21" x2="9" y2="9"/></svg></div><div class="usage-text">Press <kbd>Ctrl</kbd>+<kbd>Shift</kbd>+<kbd>P</kbd> to open the command palette</div></div><div class="usage-step"><div class="usage-icon"><svg width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><circle cx="11" cy="11" r="8"/><line x1="21" y1="21" x2="16.65" y2="16.65"/></svg></div><div class="usage-text">Type "GitIngest: Analyze Repository" and press Enter</div></div></div></div></div></div></div><div class="setup-footer">${Button({ onClick: 'startAnalysis()', icon: icons.play, children: 'Start Analysis' })}</div></div></div></div><script>const vscode=acquireVsCodeApi();function copyToClipboard(text){navigator.clipboard.writeText(text);vscode.postMessage({command:'copy',text})}function startAnalysis(){vscode.postMessage({command:'analyze'})}</script>`;
+    const styles =
+        getBaseStyles(THEME) +
+        `.setup-guide{max-width:800px;margin:0 auto;}.setup-header{text-align:center;margin-bottom:3rem;}.setup-icon{width:64px;height:64px;margin:0 auto 1rem;color:#1a1a1a;animation:pulse 2s infinite;}@keyframes pulse{0%{transform:scale(1);}50%{transform:scale(1.05);}100%{transform:scale(1);}}.setup-title{font-size:2.5rem;font-weight:700;color:#1a1a1a;margin-bottom:.5rem;}.setup-subtitle{font-size:1.1rem;color:#666;}.setup-steps{display:flex;flex-direction:column;gap:2rem;}.setup-step{display:flex;gap:1.5rem;padding:1.5rem;background:#fff4da;border:3px solid #1a1a1a;border-radius:12px;position:relative;}.setup-step::before{content:'';position:absolute;inset:0;background:#1a1a1a;border-radius:12px;transform:translate(4px,4px);z-index:-1;}.step-number{width:40px;height:40px;background:#ffc480;border:3px solid #1a1a1a;border-radius:50%;display:flex;align-items:center;justify-content:center;font-weight:700;font-size:1.2rem;flex-shrink:0;}.step-content{flex-grow:1;}.step-title{font-size:1.5rem;font-weight:600;margin-bottom:1rem;color:#1a1a1a;margin-top:0;}.step-description p{margin-bottom:1rem;color:#1a1a1a;}.step-card{background:#fff;border:3px solid #1a1a1a;border-radius:8px;padding:1rem;}.card-item{display:flex;align-items:center;gap:1rem;padding:.75rem;border-bottom:2px solid #f0f0f0;}.card-item:last-child{border-bottom:none;}.card-icon{color:#1a1a1a;height:20px;}.card-text{font-size:1rem;color:#1a1a1a;}.card-link{color:#0366d6;text-decoration:none;font-weight:500;}.card-link:hover{text-decoration:underline;}.code-block-wrapper{margin:1rem 0;}.code-block{background:#1a1a1a;border-radius:8px;padding:1rem;margin-bottom:.5rem;display:flex;justify-content:space-between;align-items:center;}.code-block code{color:#fff;font-family:'Menlo','Monaco','Courier New',monospace;}.code-note{font-size:.9rem;color:#666;margin:.5rem 0;}.copy-button{background:transparent;border:none;color:#fff;cursor:pointer;padding:.25rem;border-radius:4px;transition:background-color .2s;}.copy-button:hover{background:rgba(255,255,255,.1);}.inline-code{background:#1a1a1a;color:#fff;padding:.2rem .4rem;border-radius:4px;font-family:'Menlo','Monaco','Courier New',monospace;font-size:.9rem;}.usage-steps{display:flex;flex-direction:column;gap:1rem;}.usage-step{display:flex;align-items:center;gap:1rem;padding:1rem;background:#fff;border:2px solid #1a1a1a;border-radius:8px;}.usage-icon{color:#1a1a1a;height:20px;}.usage-text{color:#1a1a1a;font-size:1rem;}kbd{background:#1a1a1a;color:#fff;padding:.2rem .4rem;border-radius:4px;font-family:'Menlo','Monaco','Courier New',monospace;font-size:.9rem;margin:0 .2rem;}.setup-footer{margin-top:3rem;text-align:center;}.setup-footer button{font-size:1.1rem;padding:1rem 2rem;}@media (max-width:640px){.setup-step{flex-direction:column;}.step-number{margin-bottom:1rem;}}`;
+    return createHtmlDocument(content, styles);
+}
+
+// Utils
+
+export function createHtmlDocument(content: string, styles: string): string {
+    return `<!DOCTYPE html><html><head><style>${styles}</style></head><body>${content}<script>const vscode = acquireVsCodeApi();</script></body></html>`;
+}
