--- src/webview/templates/index.ts
@@ -1,4 +1,0 @@
-export { getErrorContent } from './errorTemplate';
-export { getLoadingContent } from './loadingTemplate';
-export { getResultsContent } from './resultsTemplate';
-export { getSetupGuideContent } from './setupTemplate';
