--- src/webview/utils/statusIcons.ts
@@ -1,14 +1,0 @@
-import { icons } from './icons';
-
-export function getStatusIcon(type: string): string {
-    switch (type) {
-        case 'success':
-            return `<svg width="20" height="20" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round">${icons.success}</svg>`;
-        case 'error':
-            return `<svg width="20" height="20" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round">${icons.error}</svg>`;
-        case 'warning':
-            return `<svg width="20" height="20" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round">${icons.warning}</svg>`;
-        default:
-            return `<svg width="20" height="20" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round">${icons.info}</svg>`;
-    }
-}
